+++ packages/terra-framework-site/src/Index.jsx
@@ -1,8 +1,18 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
+import { HashRouter as Router } from 'react-router-dom';
+import { Provider } from 'xfc';
+
 import App from './app/App';
 import { routes, navigation } from './appConfig';
 
+Provider.init({
+  acls: ['*'],
+  secret: () => (Promise.resolve('Success')),
+});
+
 ReactDOM.render((
-  <App routeConfig={routes} navigation={navigation} />
+  <Router>
+    <App routeConfig={routes} navigation={navigation} />
+  </Router>
 ), document.getElementById('root'));