+++ packages/terra-layout/examples/index-examples/LayoutNoHeader.jsx
@@ -1,15 +1,13 @@
 import React from 'react';
-import Layout from 'terra-layout';
+import Layout from '../../lib/Layout';
 
 import ContentExample from './ContentExample';
 import MenuExample from './MenuExample';
-import ToolbarExample from './ToolbarExample';
 
 const layout = () => (
   <Layout
-    header={<ToolbarExample />}
     menu={<MenuExample />}
-    menuText="Menu Text That Is Incredibly Long To Demonstrate Truncation"
+    menuText="Menu"
     style={{ height: '400px', width: '100%' }}
   >
     <ContentExample />