+++ packages/terra-navigation-layout/CHANGELOG.md
@@ -4,6 +4,9 @@ Changelog
 Unreleased
 ----------
 
+###Changed
+* Updating tests to point to new terra-framework-site location.
+
 1.0.0 - (December 5, 2017)
 ------------------
 * Initial release of terra-navigation-layout